post controller and mongo transactions implemented

+++ src/main/java/com/barca/blogmanager/services/internal/PostServiceImpl.java
@@ -6,6 +6,8 @@ import com.barca.blogmanager.models.Post;
 import com.barca.blogmanager.repositories.PostRepository;
 import com.barca.blogmanager.services.PostService;
 import lombok.RequiredArgsConstructor;
+
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -18,10 +20,10 @@ public class PostServiceImpl implements PostService {
   private final PostRepository postRepository;
 
   @Override
-  public void createPost(PostCreationDto postDto) {
+  public Post createPost(String userId, String userName, PostCreationDto postDto) {
 
-    Post post = new Post(null, postDto.title(), postDto.description(), postDto.content(), null, 0);
-    postRepository.save(post);
+    Post post = new Post(null, userId, userName, postDto.title(), postDto.content(), null, 0);
+    return postRepository.save(post);
   }
 
   @Override
@@ -35,12 +37,16 @@ public class PostServiceImpl implements PostService {
   }
 
   @Override
-  public void deletePost(String id) {
+  public void deletePost(String userId, String postId) {
 
     // TODO implement logic to delete associated comments
 
-    Optional<Post> result = postRepository.findById(id);
+    Optional<Post> result = postRepository.findById(postId);
     Post post = result.orElseThrow();
+
+    if (!(userId.equals(post.getUserId()))) {
+      throw new DataIntegrityViolationException("Invalid user ID");
+    }
 
     postRepository.deleteById(post.getId());
   }
